+++ config/dependencies.global.php
@@ -1,6 +1,10 @@
 <?php
 
+use ZfrEbWorker\Cli\PublisherCommand;
+use ZfrEbWorker\Cli\WorkerCommand;
+use ZfrEbWorker\Container\PublisherCommandFactory;
 use ZfrEbWorker\Container\QueuePublisherFactory;
+use ZfrEbWorker\Container\WorkerCommandFactory;
 use ZfrEbWorker\Container\WorkerMiddlewareFactory;
 use ZfrEbWorker\Publisher\QueuePublisher;
 use ZfrEbWorker\Middleware\WorkerMiddleware;
@@ -13,22 +17,10 @@ return [
         ],
 
         'factories' => [
+            PublisherCommand::class => PublisherCommandFactory::class,
             QueuePublisher::class   => QueuePublisherFactory::class,
-            WorkerMiddleware::class => WorkerMiddlewareFactory::class
+            WorkerCommand::class    => WorkerCommandFactory::class,
+            WorkerMiddleware::class => WorkerMiddlewareFactory::class,
         ]
     ],
-
-    'zfr_eb_worker' => [
-        /**
-         * Array of queue name => queue URL
-         */
-
-        'queues' => [],
-
-        /**
-         * Array of messages names => middlewares to execute
-         */
-
-        'messages' => []
-    ]
 ];